Fixed bug with unlimited holding being too large
Former-commit-id: 150e9dcdda69fc2a8f4bbf2c23f2d54a7703197a

+++ exporters/field_exporter/FieldExporter/Components/GamepiecePropertiesForm.cs
@@ -33,7 +33,7 @@ namespace FieldExporter.Components
             if (gamepieceCheckBox.Checked)
             {
                 if (holdingLimitCheckBox.Checked)
-                    return new Exporter.Gamepiece(id, Spawnpoint, (uint)holdingLimitUpDown.Value);
+                    return new Exporter.Gamepiece(id, Spawnpoint, (ushort)holdingLimitUpDown.Value);
                 else
                     return new Exporter.Gamepiece(id, Spawnpoint);
             }
